Update README.md
This is an image of me at my grandparent's home. As a child, this room was my favorite. It was often "off-limits" because it was meant for guests, but I was always drawn to the red couch, red rug, and giant gold mirror. Initially, I was planning to make the mirror smaller and outline the sides with the scalloping. I wanted to utilize the looping method, but I found myself stuck on how to end the loop. As a result, I decided to bring the image view closer to highlight how large the mirror is. When completing this assignment, I discovered the hidden color feature in "tools," which was very helpful. Using Slack was also useful when I ran into the looping problem because even watching many YouTube videos, I found that getting help from my peers was most successful.